Address #186 review: int64 progress deltas; document non-negative contract
- Keep the flusher's deltas as int64 end-to-end and widen AddProgress to
  int64 params, removing the narrowing int64->int cast (would wrap on 32-bit
  for absurd run counts) flagged by review.
- Document that AddProgress deltas must be non-negative; the executor's
  source counters are increment-only, so no runtime guard is warranted.

+++ internal/backfill/store.go
@@ -116,7 +116,10 @@ func (s *Store) Complete(id uuid.UUID, failed bool) error {
 // failed_runs write per run, all on the same backfills row, which serialize
 // through dqlite's single writer and starve concurrent control-plane writes
 // (e.g. cancellation) past the busy-retry budget.
-func (s *Store) AddProgress(id uuid.UUID, completedDelta, failedDelta int) error {
+//
+// completedDelta and failedDelta are the counts to add and must be
+// non-negative; the executor's source counters are increment-only.
+func (s *Store) AddProgress(id uuid.UUID, completedDelta, failedDelta int64) error {
 	if completedDelta == 0 && failedDelta == 0 {
 		return nil
 	}
